# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Persona Debate Local API
  version: 1.0.0
  description: |
    Local-first Node.js API for persona/debate/chat/governance workflows.
    All responses use a standard envelope:
    - success: { ok: true, data: ... }
    - error: { ok: false, error: { code, message, details } }
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
  - cookieAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/personas:
    get:
      tags: [Personas]
      summary: List personas
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: tag
          schema: { type: string }
      responses:
        '200':
          description: Personas list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
    post:
      tags: [Personas]
      summary: Create persona
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Persona' }
            examples:
              sample:
                value:
                  id: frank-bot
                  displayName: Franck
                  role: Planner
                  description: Event-planning specialist persona
                  systemPrompt: You are Franck, practical and expressive.
                  speakingStyle:
                    tone: friendly
                    verbosity: concise
                    quirks: [theatrical emphasis]
                  expertiseTags: [event planning, aesthetics]
                  biasValues: [quality-first]
                  debateBehavior: Steelman first, then challenge assumptions.
                  knowledgePackIds: []
      responses:
        '201':
          description: Persona created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /api/personas/{id}:
    get:
      tags: [Personas]
      summary: Get persona
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Persona
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    put:
      tags: [Personas]
      summary: Update persona
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Persona' }
      responses:
        '200':
          description: Persona updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      tags: [Personas]
      summary: Delete persona
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Persona deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
...
  /api/knowledge:
    get:
      tags: [Knowledge]
      summary: List knowledge packs
      responses:
        '200':
          description: Knowledge packs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
    post:
      tags: [Knowledge]
      summary: Create knowledge pack
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KnowledgePack' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/knowledge/{id}:
    get:
      tags: [Knowledge]
      summary: Get knowledge pack
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Knowledge pack
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
    put:
      tags: [Knowledge]
      summary: Update knowledge pack
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KnowledgePack' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
    delete:
      tags: [Knowledge]
      summary: Delete knowledge pack
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/knowledge/ingest:
    post:
      tags: [Knowledge]
      summary: Ingest uploaded file into knowledge pack
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                id: { type: string }
                title: { type: string }
                description: { type: string }
                tags:
                  description: Optional comma-separated tags
                  type: string
      responses:
        '201':
          description: Ingested
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/knowledge/ingest-url:
    post:
      tags: [Knowledge]
      summary: Ingest a URL into a knowledge pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string }
                id: { type: string }
                title: { type: string }
                description: { type: string }
                tags:
                  description: Optional comma-separated tags or array
                  oneOf:
                    - type: string
                    - type: array
                      items: { type: string }
                mode:
                  type: string
                  description: create, append, or overwrite
                summarize:
                  type: boolean
      responses:
        '201':
          description: Ingested
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/knowledge/preview-url:
    post:
      tags: [Knowledge]
      summary: Preview a URL before ingest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string }
                maxChars: { type: integer }
      responses:
        '200':
          description: Preview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
...
  /api/debates/{debateId}/chat:
    post:
      tags: [Debates]
      summary: Ask follow-up question grounded in transcript excerpts
      parameters:
        - in: path
          name: debateId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string }
                history:
                  type: array
                  items:
                    type: object
                    properties:
                      role: { type: string, enum: [user, assistant] }
                      content: { type: string }
                model: { type: string, default: gpt-4.1-mini }
                temperature: { type: number, default: 0.3 }
      responses:
        '200':
          description: Answer with citations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
    get:
      tags: [Debates]
      summary: Get persisted follow-up chat history for debate
      parameters:
        - in: path
          name: debateId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: History
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/persona-chats:
    post:
      tags: [Persona Chats]
      summary: Create persona collaboration chat
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePersonaChatRequest' }
            examples:
              savedPersona:
                value:
                  title: Team Chat
                  context: Planning discussion
                  knowledgePackIds: []
                  selectedPersonas:
                    - type: saved
                      id: big-tex
                  settings:
                    model: gpt-4.1-mini
                    temperature: 0.6
                    maxWordsPerTurn: 140
                    engagementMode: chat
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
    get:
      tags: [Persona Chats]
      summary: List persona chats
      responses:
        '200':
          description: Chat list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/persona-chats/{chatId}:
    get:
      tags: [Persona Chats]
      summary: Get persona chat session + messages
      parameters:
        - in: path
          name: chatId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Chat detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/persona-chats/{chatId}/messages:
    post:
      tags: [Persona Chats]
      summary: Send message to persona collaboration chat
      parameters:
        - in: path
          name: chatId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
                historyLimit: { type: integer, minimum: 4, maximum: 40, default: 14 }
                forceImage: { type: boolean, default: false }
      responses:
        '200':
          description: Orchestrated persona responses
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
...
  /api/admin/chats:
    get:
      tags: [Admin]
      summary: Chat analytics overview
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200': { description: Chat analytics, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
  /api/admin/chats/{chatId}:
    get:
      tags: [Admin]
      summary: Chat analytics detail
      parameters:
        - in: path
          name: chatId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Chat detail, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponse' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: pd_session
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        data:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details:
              nullable: true
              oneOf:
                - type: 'null'
                - type: array
                  items:
                    type: object
                    properties:
                      path: { type: string }
                      message: { type: string }
    SpeakingStyle:
      type: object
      properties:
        tone: { type: string, default: '' }
        verbosity: { type: string, default: '' }
        quirks:
          type: array
          items: { type: string }
      required: [tone, verbosity, quirks]
    Persona:
      type: object
      required: [id, displayName, systemPrompt, speakingStyle, expertiseTags, biasValues, debateBehavior, knowledgePackIds]
      properties:
        id:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        displayName: { type: string }
        role: { type: string, default: '' }
        description: { type: string, default: '' }
        systemPrompt: { type: string }
        speakingStyle: { $ref: '#/components/schemas/SpeakingStyle' }
        expertiseTags:
          type: array
          items: { type: string }
        biasValues:
          oneOf:
            - type: array
              items: { type: string }
            - type: string
        debateBehavior: { type: string, default: '' }
        knowledgePackIds:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdHocPersona:
      allOf:
        - $ref: '#/components/schemas/Persona'
      required: [displayName, systemPrompt, speakingStyle, expertiseTags, biasValues, debateBehavior, knowledgePackIds]
      properties:
        id:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
    SelectedPersona:
      oneOf:
        - type: object
          required: [type, id]
          properties:
            type: { type: string, enum: [saved] }
            id: { type: string }
        - type: object
          required: [type, persona]
          properties:
            type: { type: string, enum: [adhoc] }
            savePersona: { type: boolean, default: false }
            persona: { $ref: '#/components/schemas/AdHocPersona' }
...
    PersonaChatSettings:
      type: object
      properties:
        model: { type: string, default: gpt-4.1-mini }
        temperature: { type: number, minimum: 0, maximum: 2, default: 0.6 }
        maxWordsPerTurn: { type: integer, minimum: 40, maximum: 400, default: 140 }
        engagementMode: { type: string, enum: [chat, panel, debate-work-order], default: chat }
    CreatePersonaChatRequest:
      type: object
      required: [selectedPersonas]
      properties:
        title: { type: string, default: Persona Collaboration Chat }
        context: { type: string, default: '' }
        knowledgePackIds:
          type: array
          items: { type: string }
          default: []
        selectedPersonas:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SelectedPersona' }
        settings: { $ref: '#/components/schemas/PersonaChatSettings' }
...
    KnowledgePack:
      type: object
      required: [id, title, content]
      properties:
        id:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        title: { type: string }
        description: { type: string, default: '' }
        tags:
          type: array
          items: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'